Week6_Lab/requirements.txt
groq>=0.11.0
snowflake-connector-python>=3.6.0
python-dotenv>=1.0.0
streamlit>=1.32.0
pandas>=2.0.0
altair>=5.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
